requirements.txt
django

moore

django-debug-toolbar # 网页调试包，在config中定义debug_toolbar的配置，可以在html中显示调试工具
django-environ
django-extensions
django-filter
django-flat-responsive
django-redis-cache
djangorestframework==3.6.2
djangorestframework-jwt

hiredis

psycopg2

requests

pyjwt
